2-hop 구현을 위한 type, adapter 리펙토링
