+++ apps/frontend/components/LawGraphPanel.tsx
@@ -1,10 +1,9 @@
 import { useEffect, useRef, useState } from "react";
 import { Network as NetworkIcon, Loader2, AlertCircle, BarChart3 } from "lucide-react";
-import { DataSet } from "vis-data";
 import { Network } from "vis-network";
 import type { Options } from "vis-network";
 import { queryGraph } from "@/lib/api-client";
-import { toGraphData } from "@/lib/graph-adapter";
+import { buildVisDatasets, toGraphData } from "@/lib/graph-adapter";
 import type { GraphNode, LawGraphData } from "@/lib/graph-types";
 
 interface LawGraphPanelProps {
@@ -49,49 +48,6 @@ const VIS_OPTIONS: Options = {
   },
 };
 
-function lawNodeColor(lawType: string | undefined) {
-  switch (lawType) {
-    case "시행령":
-      return { background: "hsl(217, 75%, 86%)", border: "hsl(217, 55%, 65%)", highlight: { background: "hsl(217, 75%, 79%)", border: "hsl(217, 55%, 55%)" } };
-    case "시행규칙":
-      return { background: "hsl(217, 55%, 82%)", border: "hsl(217, 40%, 62%)", highlight: { background: "hsl(217, 55%, 75%)", border: "hsl(217, 40%, 52%)" } };
-    case "법":
-    default:
-      return { background: "hsl(217, 91%, 92%)", border: "hsl(217, 60%, 70%)", highlight: { background: "hsl(217, 91%, 85%)", border: "hsl(217, 60%, 60%)" } };
-  }
-}
-
-function buildVisDatasets(graphData: LawGraphData) {
-  const visNodes = graphData.nodes.map((n) => ({
-    id: n.id,
-    label: n.label,
-    size: n.isCenter ? 28 : 18,
-    color: n.isCenter
-      ? { background: "hsl(217, 91%, 50%)", border: "hsl(217, 91%, 40%)", highlight: { background: "hsl(217, 91%, 45%)", border: "hsl(217, 91%, 35%)" } }
-      : n.kind === "law"
-        ? lawNodeColor(n.lawType)
-        : { background: "hsl(142, 71%, 93%)", border: "hsl(142, 60%, 65%)", highlight: { background: "hsl(142, 71%, 86%)", border: "hsl(142, 60%, 55%)" } },
-    font: { color: n.isCenter ? "hsl(217, 91%, 30%)" : "hsl(220, 30%, 20%)", size: n.isCenter ? 13 : 12 },
-  }));
-
-  const visEdges = graphData.edges.map((e) => ({
-    id: e.id,
-    from: e.source,
-    to: e.target,
-    label: e.relationType === "child_law"
-      ? "하위"
-      : e.relationType === "delegation"
-        ? "위임"
-        : e.relationType === "reference"
-          ? (e.paragraphNos?.length ? `참조 · 제${parseInt(e.paragraphNos[0], 10) || e.paragraphNos[0]}항` : "참조")
-          : "",
-  }));
-
-  return {
-    nodes: new DataSet(visNodes),
-    edges: new DataSet(visEdges),
-  };
-}
 
 export function LawGraphPanel({ lastQuery, queryKey, isActive, onNodeSelect, onGraphDataChange }: LawGraphPanelProps) {
   const containerRef = useRef<HTMLDivElement>(null);
