+++ apps/frontend/lib/graph-adapter.ts
@@ -1,4 +1,75 @@
+import { DataSet } from "vis-data";
 import type { GraphEdge, GraphEdgeRelationType, GraphNode, GraphQueryResponse, LawGraphData } from "./graph-types";
+
+// ── vis-network 색상 헬퍼 ─────────────────────────────────────────────────────
+
+function lawNodeColor(lawType: string | undefined) {
+  switch (lawType) {
+    case "시행령":
+      return { background: "hsl(217, 75%, 86%)", border: "hsl(217, 55%, 65%)", highlight: { background: "hsl(217, 75%, 79%)", border: "hsl(217, 55%, 55%)" } };
+    case "시행규칙":
+      return { background: "hsl(217, 55%, 82%)", border: "hsl(217, 40%, 62%)", highlight: { background: "hsl(217, 55%, 75%)", border: "hsl(217, 40%, 52%)" } };
+    case "법":
+    default:
+      return { background: "hsl(217, 91%, 92%)", border: "hsl(217, 60%, 70%)", highlight: { background: "hsl(217, 91%, 85%)", border: "hsl(217, 60%, 60%)" } };
+  }
+}
+
+export function buildVisDatasets(graphData: LawGraphData) {
+  const visNodes = graphData.nodes.map((n) => ({
+    id: n.id,
+    label: n.label,
+    size: n.isCenter ? 28 : 18,
+    color: n.isCenter
+      ? { background: "hsl(217, 91%, 50%)", border: "hsl(217, 91%, 40%)", highlight: { background: "hsl(217, 91%, 45%)", border: "hsl(217, 91%, 35%)" } }
+      : n.kind === "law"
+        ? lawNodeColor(n.lawType)
+        : { background: "hsl(142, 71%, 93%)", border: "hsl(142, 60%, 65%)", highlight: { background: "hsl(142, 71%, 86%)", border: "hsl(142, 60%, 55%)" } },
+    font: { color: n.isCenter ? "hsl(217, 91%, 30%)" : "hsl(220, 30%, 20%)", size: n.isCenter ? 13 : 12 },
+  }));
+
+  const visEdges = graphData.edges.map((e) => ({
+    id: e.id,
+    from: e.source,
+    to: e.target,
+    label: e.relationType === "child_law"
+      ? "하위"
+      : e.relationType === "delegation"
+        ? "위임"
+        : e.relationType === "reference"
+          ? (e.paragraphNos?.length ? `참조 · 제${parseInt(e.paragraphNos[0], 10) || e.paragraphNos[0]}항` : "참조")
+          : "",
+  }));
+
+  return {
+    nodes: new DataSet(visNodes),
+    edges: new DataSet(visEdges),
+  };
+}
+
+// ── 그래프 병합 (expand 시 증분 추가) ─────────────────────────────────────────
+
+export function mergeGraphData(
+  existing: LawGraphData,
+  newNodes: GraphNode[],
+  newEdges: GraphEdge[],
+  sourceHop: number,
+): LawGraphData {
+  const seenNodes = new Set(existing.nodes.map((n) => n.id));
+  const seenEdges = new Set(existing.edges.map((e) => e.id));
+
+  const addedNodes = newNodes
+    .filter((n) => !seenNodes.has(n.id))
+    .map((n) => ({ ...n, hop: sourceHop + 1 }));
+
+  const addedEdges = newEdges.filter((e) => !seenEdges.has(e.id));
+
+  return {
+    ...existing,
+    nodes: [...existing.nodes, ...addedNodes],
+    edges: [...existing.edges, ...addedEdges],
+  };
+}
 
 export function toGraphData(resp: GraphQueryResponse): LawGraphData | null {
   if (!resp.law_name || !resp.results.length) return null;
@@ -18,6 +89,7 @@ export function toGraphData(resp: GraphQueryResponse): LawGraphData | null {
         lawName: resp.law_name,
         articleNo: resp.article_no,
         isCenter: true,
+        hop: 0,
       }
     : {
         id: centerId,
@@ -25,6 +97,7 @@ export function toGraphData(resp: GraphQueryResponse): LawGraphData | null {
         kind: "law",
         lawName: resp.law_name,
         isCenter: true,
+        hop: 0,
       };
 
   const nodes: GraphNode[] = [centerNode];
